Change score to float

+++ db/migrate/20211222130146_add_average_score_to_stake_pools.rb
@@ -1,5 +1,5 @@
 class AddAverageScoreToStakePools < ActiveRecord::Migration[6.1]
   def change
-    add_column :stake_pools, :average_score, :integer
+    add_column :stake_pools, :average_score, :float
   end
 end
